Alterando painel de exibir valores na classe de dois jogadores

+++ Complexos/Fichas_de_Poker.java
@@ -8,37 +8,37 @@ public class Fichas_de_Poker{
         Scanner teclado = new Scanner(System.in);
 
         boolean loop = true;
-        int valorApostado;
+        float valorApostado;
 
         // Variaveis de Fichas:
 
         //j1
-        int j1_100;
-        int j1_50;
-        int j1_25;
-        int j1_10;
-        int j1_5;
+        float j1_100;
+        float j1_50;
+        float j1_25;
+        float j1_10;
+        float j1_5;
 
         //j2
-        int j2_100;
-        int j2_50;
-        int j2_25;
-        int j2_10;
-        int j2_5;
+        float j2_100;
+        float j2_50;
+        float j2_25;
+        float j2_10;
+        float j2_5;
 
         //j3
-        int j3_100;
-        int j3_50;
-        int j3_25;
-        int j3_10;
-        int j3_5;
+        float j3_100;
+        float j3_50;
+        float j3_25;
+        float j3_10;
+        float j3_5;
 
         //j4
-        int j4_100;
-        int j4_50;
-        int j4_25;
-        int j4_10;
-        int j4_5;
+        float j4_100;
+        float j4_50;
+        float j4_25;
+        float j4_10;
+        float j4_5;
 
         while (loop) {
 
@@ -56,33 +56,33 @@ public class Fichas_de_Poker{
             case 1:
 
                 System.out.printf("Valor apostado: R$");
-                valorApostado = teclado.nextInt();
+                valorApostado = teclado.nextFloat();
 
                 System.out.println("Digite a quantidade de fichas de ambos os jogadores\n");
 
                 System.out.println("---------- Jogador 1 ----------");
                 System.out.printf("Fichas de 100: ");
-                j1_100 = teclado.nextInt();
+                j1_100 = teclado.nextFloat();
                 System.out.printf("Fichas de 50: ");
-                j1_50 = teclado.nextInt();
+                j1_50 = teclado.nextFloat();
                 System.out.printf("Fichas de 25: ");
-                j1_25 = teclado.nextInt();
+                j1_25 = teclado.nextFloat();
                 System.out.printf("Fichas de 10: ");
-                j1_10 = teclado.nextInt();
+                j1_10 = teclado.nextFloat();
                 System.out.printf("fichas de 5: ");
-                j1_5 = teclado.nextInt();
+                j1_5 = teclado.nextFloat();
 
                 System.out.println("---------- Jogador 2 ----------");
                 System.out.printf("Fichas de 100: ");
-                j2_100 = teclado.nextInt();
+                j2_100 = teclado.nextFloat();
                 System.out.printf("Fichas de 50: ");
-                j2_50 = teclado.nextInt();
+                j2_50 = teclado.nextFloat();
                 System.out.printf("Fichas de 25: ");
-                j2_25 = teclado.nextInt();
+                j2_25 = teclado.nextFloat();
                 System.out.printf("Fichas de 10: ");
-                j2_10 = teclado.nextInt();
+                j2_10 = teclado.nextFloat();
                 System.out.printf("fichas de 5: ");
-                j2_5 = teclado.nextInt();
+                j2_5 = teclado.nextFloat();
                 
                 DoisJogadores opcao1 = new DoisJogadores(j1_100, j1_50, j1_25, j1_10, j1_5, j2_100, j2_50, j2_25, j2_10, j2_5, valorApostado);
